Build product.json to display excluded nodes
Lines 40-47 will build and output a product.json, intended to show the exclude nodes following preprocessing on brainlife.io following app completion.

Lines 40-43 build the product.json, which is a simple json dictionary with a key of "brainlife" as an empty list, with the message inside a dictionary appended. Currently, the node labels are added to the "here are the excluded nodes" string with a comma and space separator between the nodes. But this, and the preceding message, can be however you want it to be formatted as long as it will work with the string python datatype.

The remaining lines save it out to the current working directory.